+++ src/roq/samples/example-1/main.cpp
@@ -2,10 +2,10 @@
 
 #include <gflags/gflags.h>
 
-#include "roq/application.h"
-#include "roq/client.h"
-
 #include "roq/logging.h"
+#include "roq/service.h"
+
+#include "roq/client.h"
 
 // command line options
 
@@ -162,12 +162,12 @@ class Strategy final : public client::Handler {
   client::Dispatcher& _dispatcher;
 };
 
-// Application is a base class used to initialize the environment
+// Service is a base class used to initialize the environment
 // it is not required, only for your convenience
 
-class Controller final : public Application {
+class Controller final : public Service {
  public:
-  using Application::Application;
+  using Service::Service;
 
  protected:
   int main_helper(const roq::span<std::string_view>& args) {