+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Replaced `Application` with `Service`
+
 ## 0.4.4 &ndash; 2020-09-20
 
 ### Added